refact: refactor mysql_password resource + add tests
